ai/requirements.txt
absl-py==2.3.1
alembic==1.16.4
annotated-types==0.7.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
filetype==1.2.0
fonttools==4.59.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
grpcio==1.73.1
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.8
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
optuna==4.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pillow-avif-plugin==1.5.2
pillow_heif==1.0.0
platformdirs==4.3.8
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
roboflow==1.2.1
scipy==1.16.0
sentry-sdk==2.33.1
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.41
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.168
ultralytics-thop==2.0.14
urllib3==2.5.0
wandb==0.21.0
Werkzeug==3.1.3
